requirements.txt
openai
google-genai

python-dotenv
pydantic
pandas

langchain
langchain-google-genai
langchain-openai
langchain-ollama
langchain-tavily

tavily-python
